+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/other/utils/TwoWDRobotWithSpinnerAndLift.java
@@ -22,8 +22,8 @@ public class TwoWDRobotWithSpinnerAndLift extends GamepadExtended {
 
         List<String> motorNames = new ArrayList<String>();
 
-        motorNames.add(hardwareMap.appContext.getString(R.string.LEFT_DRIVE_1));
-        motorNames.add(hardwareMap.appContext.getString(R.string.RIGHT_DRIVE_1));
+        motorNames.add(hardwareMap.appContext.getString(R.string.DRIVETRAIN_LEFT_DRIVE_1));
+        motorNames.add(hardwareMap.appContext.getString(R.string.DRIVETRAIN_RIGHT_DRIVE_1));
 
         this.drivetrainManager4WD = new DrivetrainManager4WD(motorNames, hardwareMap);
